+++ src/template.c
@@ -1,6 +1,6 @@
 #include "include/template.h"
 
-template_T* init_template(unsigned char* code, u_int64_t id)
+template_T* init_template(unsigned char* code, u_int32_t id)
 {
     template_T* temp = calloc(1, sizeof(struct TEMPLATE_STRUCT));
     temp->code = code;
